backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
anthropic==0.43.0
pydantic==2.10.4
python-dotenv==1.0.1
# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
# Configuration
pydantic-settings>=2.7.0
# Logging
structlog>=24.4.0
# Geocoding
pgeocode>=0.4.0
# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
